wip 1.0 works fine should refactor request

+++ src/responses/Response.hpp
@@ -30,7 +30,7 @@ public:
 	void			generateDirectoryListing(const std::string& directoryPath);
 	void			generateCGIResponse(const std::string &cgi_response);
 	const char*		getContent();
-	int				getContentLength();
+	ssize_t			getContentLength();
 	// const char* toCString();
 
 	/* Exceptions */
